UI Skeleton Branch Created
This includes an implementation of the Conversatio Activity mechanics.
Placeholder messages have been inserted where features are to be
created, but the mechanics are there. Creating Conversations on the
list, clicking, and longclicking them is supported.

The Messaging Activity will be very similar to this one with a few
exceptions. The class holding the message info will be a little more
complex, with a special graphic background, and there will be only one
EditText.