V1 Complete
This skeleton UI is fully-ready to have the proper messaging and login
logic linked into it. It has some shortcomings that can be improved when
we have time, but for sake of priorities, let's focus on getting it in
shape first.

NEXT STEPS:
-Insert Sync and Push AZURE logic into Conversation and Messaging
activities
-Code proper Timestamp format that will work with AZURE SQL Lite table
(Use GregorianCalendar class for this)
-Link Login Activity to the Conversation Activity on successful login

Upon completion of these steps, we will have a functional messenger.

SECURITY STEPS:
-Implement key exchange and store credentials
-Implement Android Keystore to hold credentials used to decrypt
-Implement Pepper to further-protect login credentials
-Ensure that Azure connection credentials are secured and not open to
decompile and theft
-Implement Handle generation (the Settings menu in the Conversation
Activity can be used for this) and Anonymous conversation feature
-Write hash of text sent and text received to verify integrity

QUALITY FEATURES (Nice to have but best saved for last)
-Fix listview so that the text fields don't overlap list items
-Set messages to appear in easy-to-read bubbles
-Create custom app icon
-Set custom color scheme

AMBITIOUS STEPS (If we get enough time to really reach)
-Implement a Notification system when user gets new messages
-Implement GCM and Event Hubs on Azure (can be used to do the above)
-Create local storage for messages to be retained asynchronously
-Have messages deleted upon being sent