Working Messenger V1.1
Login logic has been altered to use script calls to the azure server
instead of syncing the user table data directly. By using semaphores on
the thread executing login, the login mehtod first fetches the salt
associated with the username, then performs the hash with the salt and
pepper on the input password and sends it to the server. The server
sends back a yes or no. This feature:
A) Prevents the app from being aware of user data in its entirety
(especially the password)
B) Makes it POSSIBLE to create security keys, though I cannot find a way
of definitively doing so. It may simply be beyond the scope of this
project, but we should include that as a Next Steps in the report if
not.